feat(global): добавить snackbar в App (#61)
Jira issue: VEGA-604
