+++ src/providers/snackbar/Snackbar.css
@@ -1,0 +1,15 @@
+.snackbar {
+  width: 400px;
+}
+
+.snackbar-portal {
+  position: fixed;
+  z-index: 100;
+
+  bottom: 0;
+
+  display: flex;
+  justify-content: center;
+
+  width: 100%;
+}
