+++ src/providers/snackbar/index.ts
@@ -1,0 +1,1 @@
+export { SnackbarProvider, useSnackbar } from './Snackbar';
